Add time spent histogram, move getDurationText to /const
