+++ src/models/stats.js
@@ -19,6 +19,7 @@ export default class StatsModel {
 
     this.moneyInfo = null;
     this.transportInfo = null;
+    this.timeInfo = null;
 
     this.countStats();
   }
@@ -49,8 +50,25 @@ export default class StatsModel {
     return prepareStatsData(typesWithCounts);
   }
 
+  _createTimeInfo() {
+    const types = this._typesGroups.reduce((acc, typesGroup) => acc.concat(typesGroup.types), []);
+
+    const typesWithTimes = types.map((type) => {
+      const cardsByType = getCardsByType(this._cards, type);
+      const durationSum = countSumByField(cardsByType, `duration`);
+
+      return [
+        type,
+        durationSum
+      ];
+    });
+
+    return prepareStatsData(typesWithTimes);
+  }
+
   countStats() {
     this.moneyInfo = this._createMoneyInfo();
     this.transportInfo = this._createTransportInfo();
+    this.timeInfo = this._createTimeInfo();
   }
 }
